Delegate creation of shared channels to a global object
In previous versions shared channels had no means of closing their
underlying channels. If two channels share the same underlying channel
the invocation of the close method in first channel affects the second
one.

However, the underlying channel can be closed if all channels that share
this channel invoked the close method. To implement this,
SharedChannelFactory is responsible for creation of all shared channels.
It stores the counter of how many shared channels use a specific
channel. If that counter reaches zero the channel closes.