Optimize the way metadata is set when writing the tree into a filesystem
As an ad-hoc solution every time before writing a file or a directory
into the parent directory the modification time was red and then was
set back again. It works well when the parent folder has few
subdirectories or files. However, if there are thousands subdirectories,
the modification data is red and set thousands times as well, instead
of doing it once when all files are written.

The newer implementation, however, requires to store all the directories
that have children and the metadata setting happens after all files and
directories of the tree are created. It doesn't increase the memory
footprint too much because only references are stored.