Replace ArchiveEntityFactory with ArchiveTreeBuilder
ArchiveEntityFactory was initially designed as a number of static
methods with different parameters to construct archive entities.
But this design would lead to an enormous size of ArchiveEntityFactory
as well as static only methods were a limitation.
The use of a common one function interface a.k.a. a lambda function
proved its efficiency: different implementations are separated into
different classes and their objects can also store additional data.