Seal the AESStrategy interface
AESStrategy is an interface to encrypt and decrypt any data passed to
its methods. In order to provide a good level of security AES is
intended to use with the variety block cipher modes implemented by
the concrete classed. This, however, creates a situation where
an open-ended number of subclasses requires different parameters each
time. So to solve it, it is necessary to limit the amount of subclasses
and then create a class to generate the necessary arguments for a given
strategy.